First draft of Create context (#3)
* finised layout generating for real screen;
TODO: click to add view;

* finished view insert list and view insert mechanism;
TODO: 1). drag to keep views; 2). improve coop on different mechanisms;

* fixed minor issues;

* improved view adding;
now views can be rescaled;
TODO: as discussed;

* improved a series interaction mechanism as discussed;

* fixed minor bugs;

* fixed bugs in interaction;
imporoved interaction as discussed;

+++ implementation/js/ide/view/create/arrows.js
@@ -2,7 +2,7 @@
 
 	app.view('Create.Arrows', {
 		template: '@view/create/arrows.html',
-		coop: ['click-endpoint'],
+		coop: ['click-endpoint', 'point-layout-reset-confirmed'],
 		lines: {//store lines attached to current point
 			up: '',
 			down: '',
@@ -15,6 +15,10 @@
 
 		},
 		onClickEndpoint: function(e){
+			//hide side menu if necesary
+			var $trigger = this.parentCt.$el.find('.side-menu-trigger');
+			if($trigger.hasClass('active'))
+				this.parentCt.toggleSideMenu($trigger);
 			//reset
 			this.cleanIndication();
 			//vars
@@ -65,13 +69,33 @@
 		},
 		actions: {
 			delete: function($self){
-				if($self.find('i').hasClass('deletable')){//deletable
-					var line = this.lines[$self.data('direction')];
-					this.deleteLine(line, $self.data('direction'));
-					this.coop('sync-local');
-				}else{//NOT deletable
-					app.notify('Cannot Delete', 'This line cannot be deleted!', 'error', {icon: 'fa fa-reddit-alien'});
+				//if parectView has generated layout, send reset event
+				if(this.parentCt.generated){
+					(new (app.get('Create.LayoutResetConfirm'))({
+						data: {
+							anchor: $self
+						}
+					})).overlay({
+						effect: false,
+						class: 'layout-reset-confirm-overlay create-overlay danger-title',
+					});
 				}
+				else
+					this.confirmDelete($self);
+			}
+		},
+		onPointLayoutResetConfirmed: function($anchor){
+			this.coop('layout-resetted');
+			this.confirmDelete($anchor);
+		},
+		confirmDelete: function($self){
+			if($self.find('i').hasClass('deletable')){//deletable
+				var line = this.lines[$self.data('direction')];
+				this.deleteLine(line, $self.data('direction'));
+
+				this.coop('sync-local');
+			}else{//NOT deletable
+				app.notify('Cannot Delete', 'This line cannot be deleted!', 'error', {icon: 'fa fa-reddit-alien'});
 			}
 		},
 		closeMenu: function(){
